+++ Core/FeedMediaCore.h
@@ -18,6 +18,6 @@
 #import "FMLog.h"
 #import "FMStation.h"
 
-#define FEED_MEDIA_CLIENT_VERSION @"4.0.0-beta2"
+#define FEED_MEDIA_CLIENT_VERSION @"4.0.0"
 
 #endif /* FeedMediaCore_h */